requirements.txt
# Core dependencies
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0
numpy>=1.24.0
scipy>=1.10.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0
h5py>=3.10.0

# AutoML and optimization
optuna>=3.5.0
ray[tune]>=2.9.0

# Monitoring and logging
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Utilities
pyyaml>=6.0
hydra-core>=1.3.0
omegaconf>=2.3.0
colorama>=0.4.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.12.0
flake8>=6.1.0
mypy>=1.8.0

# Additional utilities
psutil>=5.9.0
matplotlib>=3.8.0
seaborn>=0.13.0

